+++ plugin_proposal/include/FactoryPlugin.hpp
@@ -1,110 +1,32 @@
 #pragma once
 
-#include <stdexcept>
-#include <string>
-#include <memory>
-#include <sstream>
-#include <functional>
-#include <set>
-#include <link.h>
-#include <dlfcn.h>
-#include <fcntl.h>
-#include <unistd.h>
-#include <sys/mman.h>
-#include <sys/stat.h>
-#include <openssl/evp.h>
+#include "DllHandle.hpp"
 
+
+// This is the name of the only symbol to look up in the DLL
+#define FACTORY_GETTER_NAME get_factory__
+#define STRINGIFY(s) #s
+
+//
+// Class definition is only needed in client side.
+//
 template <typename FACT>
 class FactoryPlugin {
-    void* dl_handle;
-    std::string path;
-
-    void throw_error(const std::string& desc, const std::string& fpath)
-    {
-        using namespace std;
-        stringstream ss;
-        ss << "Cannot " << desc << " dll file \"" << fpath << '"';
-        throw runtime_error(ss.str());
-    }
-
-    std::string compute_md5(const std::string& fpath)
-    {
-        using namespace std;
-
-        int fd = ::open(fpath.c_str(), O_RDONLY);
-        if (fd < 0) throw_error("open", fpath);
-        struct stat sbuf;
-        ::fstat(fd, &sbuf);
-        void* ptr = mmap(0, sbuf.st_size, PROT_READ, MAP_SHARED, fd, 0);
-        if (ptr == nullptr) {
-            close(fd);
-            throw_error("mmap", fpath);
-        }
-        close(fd);
-
-        const size_t digest_len = 16;
-        unsigned char digest[digest_len];
-        EVP_Q_digest(NULL, "MD5", NULL, ptr, sbuf.st_size, digest, NULL);
-        munmap(ptr, sbuf.st_size);
-
-        std::string        output;
-        output.resize(digest_len * 2);
-        for (unsigned int i = 0 ; i < digest_len ; ++i)
-            std::sprintf(&output[i * 2], "%02x", digest[i]);
-        return output;
-    }
-public:
+    std::shared_ptr<DllHandle>  dll_handle;
     FACT* factory;
-
-    using WhiteList = std::set<std::string>;
-    FactoryPlugin(const std::string& path, const WhiteList& white_list = WhiteList()) : path(path), dl_handle(nullptr)
+public:
+    using WhiteList = DllHandle::WhiteList;
+    FactoryPlugin(const std::string& path, const WhiteList& white_list = WhiteList())
     {
-        using namespace std;
-        auto hash = compute_md5(path);
-        if (white_list.size() > 0) {
-            if (white_list.count(hash) == 0) {
-                using namespace std;
-                stringstream ss;
-                ss << "MD5 hash=" << hash << " for  file \"" << path << "\" is not in white list";
-                throw runtime_error(ss.str());
-            }
-        }
-        else {
-            using namespace std;
-            cerr << "MD5 hash=" << hash << " for \"" << path << "\" was found but whitelist is missing." << endl;
-        }
-        dl_handle = dlopen(path.c_str(), RTLD_NOW|RTLD_LOCAL);
-        if (dl_handle == nullptr) throw_error("dlopen", path);
-
-
-        struct ::link_map *lmp;
-        if (dlinfo(dl_handle, RTLD_DI_LINKMAP, &lmp) != 0) throw_error("dlinfo(RTLD_DI_LINKMAP)", path);
-        for ( ; lmp; lmp = lmp->l_next) {
-            using namespace std;
-            cout << "MD5: " << compute_md5(lmp->l_name)  << ' ' << lmp->l_name  << endl;
-        }
-
-
-        void* ptr = dlsym(dl_handle, "get_factory__");
-        if (ptr == nullptr) {
-            dlclose(dl_handle);
-            throw_error("locate get_factory__ symbol in ", path);
-        }
-        auto get_factory = (void* (*)())ptr;
+        dll_handle = std::make_shared<DllHandle>(path);
+        auto get_factory = (void* (*)()) dll_handle->getSymbol(STRINGIFY(FACTORY_GETTER_NAME));
         factory = (FACT*) (*get_factory)();
     }
-
-    ~FactoryPlugin()
-    {
-        // apparently, dlclose with actually not unload or even run dtors if there are 'UNIQUE' tagged objects
-        // So, don't depend on dtors in plugin code for back-to-back clean testing.
-        if (dl_handle != nullptr) {
-            if (dlclose(dl_handle) != 0) throw_error("dlclose", path);
-        }
-    }
-
     FACT* operator->() { return factory; }
 };
 
+//
+// Macros are only needed in library side to expose the 
+//
 #define _EXPOSE extern "C" __attribute__((visibility("default")))
-#define FACTORY_EXPOSE(fact) _EXPOSE void * get_factory__() { return (void*) &fact; }
+#define FACTORY_EXPOSE(fact) _EXPOSE void * FACTORY_GETTER_NAME() { return (void*) &fact; }
